Fix build after merge
