+++ teachertool/src/types/errorCode.ts
@@ -4,5 +4,8 @@ export enum ErrorCode {
     loadCatalogFailed = "loadCatalogFailed",
     getProjectTextAsync = "getProjectTextAsync",
     getProjectMetaAsync = "getProjectMetaAsync",
-    downloadTargetConfigAsync = "downloadTargetConfigAsync"
+    downloadTargetConfigAsync = "downloadTargetConfigAsync",
+    loadValidatorPlansFailed = "loadValidatorPlansFailed",
+    evalMissingCriteria = "evalMissingCriteria",
+    evalMissingPlan = "evalMissingPlan",
 }
